requirements/base.txt
aiocache==0.11.1
aiohttp==3.6.2
aiohttp-swagger==1.0.14
gunicorn==20.0.0
simple-settings==0.19.1
uvloop==0.14.0
aioredis==1.3.1
loguru==0.4.1
python-socketio==4.5.1
